feat(main)
* Clean main.rs
* New module with the backend implementations needed
