+++ src/chat/chat.rs
@@ -1,0 +1,118 @@
+use anyhow::Result;
+use clap::Parser;
+use core::fmt;
+use futures_lite::StreamExt;
+use iroh::{NodeAddr, NodeId};
+use iroh_gossip::net::{Event, GossipEvent, GossipReceiver};
+use iroh_gossip::proto::TopicId;
+use serde::{Deserialize, Serialize};
+use std::collections::HashMap;
+use std::str::FromStr;
+
+#[derive(Parser, Debug)]
+pub struct Args {
+    #[clap(short, long)]
+    pub name: Option<String>,
+
+    #[clap(short, long, default_value = "0")]
+    pub bind_port: u16,
+    #[clap(subcommand)]
+    pub command: Command,
+}
+
+#[derive(Parser, Debug)]
+pub enum Command {
+    Open,
+    Join { ticket: String },
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct Ticket {
+    pub topic: TopicId,
+    pub nodes: Vec<NodeAddr>,
+}
+
+impl Ticket {
+    pub fn from_bytes(bytes: &[u8]) -> Result<Self> {
+        serde_json::from_slice(bytes).map_err(Into::into)
+    }
+    pub fn to_bytes(&self) -> Vec<u8> {
+        serde_json::to_vec(self).expect("serde_json::to_vec is infallible")
+    }
+}
+
+impl fmt::Display for Ticket {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        let mut text = data_encoding::BASE32_NOPAD.encode(&self.to_bytes()[..]);
+        text.make_ascii_lowercase();
+        write!(f, "{}", text)
+    }
+}
+
+impl FromStr for Ticket {
+    type Err = anyhow::Error;
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        let bytes = data_encoding::BASE32_NOPAD.decode(s.to_ascii_uppercase().as_bytes())?;
+        Self::from_bytes(&bytes)
+    }
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct Message {
+    pub body: MessageBody,
+    pub nonce: [u8; 16],
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub enum MessageBody {
+    AboutMe { from: NodeId, name: String },
+    Message { from: NodeId, text: String },
+}
+
+impl Message {
+    pub fn from_bytes(bytes: &[u8]) -> Result<Self> {
+        serde_json::from_slice(bytes).map_err(Into::into)
+    }
+    pub fn new(body: MessageBody) -> Self {
+        Self {
+            body,
+            nonce: rand::random(),
+        }
+    }
+    pub fn to_vec(&self) -> Vec<u8> {
+        serde_json::to_vec(self).expect("serde_json::to_vec is infallible")
+    }
+}
+
+pub async fn subscribe_loop(mut receiver: GossipReceiver) -> Result<()> {
+    let mut names = HashMap::new();
+
+    while let Some(event) = receiver.try_next().await? {
+        if let Event::Gossip(GossipEvent::Received(msg)) = event {
+            match Message::from_bytes(&msg.content)?.body {
+                MessageBody::AboutMe { from, name } => {
+                    names.insert(from, name.clone());
+                    println!("> {} is now known as {}", from.fmt_short(), name);
+                }
+
+                MessageBody::Message { from, text } => {
+                    let name = names
+                        .get(&from)
+                        .map_or_else(|| from.fmt_short(), String::to_string);
+                    println!("{}: {}", name, text);
+                }
+            }
+        }
+    }
+    Ok(())
+}
+
+pub fn input_loop(line_tx: tokio::sync::mpsc::Sender<String>) -> Result<()> {
+    let mut buffer = String::new();
+    let stdin = std::io::stdin();
+    loop {
+        stdin.read_line(&mut buffer)?;
+        line_tx.blocking_send(buffer.clone())?;
+        buffer.clear();
+    }
+}
